# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RecruitHub API",
    "description": "API documentation for RecruitHub",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your bearer token in the format **Bearer {token}**"
      }
    },
    "schemas": {
...
      "EvaluationCriteriaDto": {
        "type": "object",
        "required": ["name", "ratingType"],
        "properties": {
          "criteriaId": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the evaluation criteria",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Name of the evaluation criteria",
            "example": "Problem Solving"
          },
          "description": {
            "type": "string",
            "description": "Description of the evaluation criteria",
            "example": "Ability to solve complex problems"
          },
          "ratingType": {
            "type": "string",
            "enum": ["percentage", "scale-5", "scale-10", "yes-no", "text"],
            "description": "Type of rating for this criteria",
            "example": "scale-5"
          },
          "isRequired": {
            "type": "boolean",
            "description": "Whether this criteria is required for evaluation",
            "default": true,
            "example": true
          }
        }
      },
      "RoundDto": {
        "type": "object",
        "required": ["roundNumber", "name", "startTime", "endTime", "evaluationCriteria"],
        "properties": {
          "roundNumber": {
            "type": "integer",
            "description": "The round number (1-5)",
            "minimum": 1,
            "maximum": 5,
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Name of the round",
            "example": "Technical Round"
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "description": "Start time of the round",
            "example": "2025-06-15T09:00:00Z"
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "description": "End time of the round",
            "example": "2025-06-15T12:00:00Z"
          },
          "evaluationCriteria": {
            "type": "array",
            "description": "List of evaluation criteria for this round",
            "items": {
              "$ref": "#/components/schemas/EvaluationCriteriaDto"
            },
            "minItems": 1
          }
        }
      },
      "CreateRoundDto": {
        "$ref": "#/components/schemas/RoundDto"
      },
      "UpdateRoundDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the round",
            "example": "Updated Technical Round"
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "description": "Start time of the round",
            "example": "2025-06-15T10:00:00Z"
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "description": "End time of the round",
            "example": "2025-06-15T13:00:00Z"
          },
          "evaluationCriteria": {
            "type": "array",
            "description": "List of evaluation criteria for this round",
            "items": {
              "$ref": "#/components/schemas/EvaluationCriteriaDto"
            },
            "minItems": 1
          }
        }
      },
      "SecondarySpocDto": {
        "type": "object",
        "required": ["spocId", "spocEmail", "spocName"],
        "properties": {
          "spocId": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the SPOC",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "spocEmail": {
            "type": "string",
            "format": "email",
            "description": "Email of the SPOC",
            "example": "secondary.spoc@example.com"
          },
          "spocName": {
            "type": "string",
            "description": "Name of the SPOC",
            "example": "Jane Smith"
          }
        }
      },
      "CreateDriveDto": {
        "type": "object",
        "required": [
          "name", 
          "collegeId", 
          "collegeName", 
          "role", 
          "practice", 
          "startDate", 
          "primarySpocId", 
          "primarySpocEmail", 
          "primarySpocName"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the drive",
            "example": "Summer 2025 Recruitment Drive"
          },
          "collegeId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the college",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "collegeName": {
            "type": "string",
            "description": "Name of the college",
            "example": "Harvard University"
          },
          "role": {
            "type": "string",
            "enum": ["Associate Engineer", "Business Analyst"],
            "description": "Role being recruited for",
            "example": "Associate Engineer"
          },
          "practice": {
            "type": "string",
            "enum": ["Application Development", "DevOps", "PMO", "BaUX"],
            "description": "Practice area",
            "example": "Application Development"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Start date of the drive",
            "example": "2025-06-15T09:00:00Z"
          },
          "primarySpocId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the primary SPOC",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "primarySpocEmail": {
            "type": "string",
            "format": "email",
            "description": "Email of the primary SPOC",
            "example": "primary.spoc@example.com"
          },
          "primarySpocName": {
            "type": "string",
            "description": "Name of the primary SPOC",
            "example": "John Doe"
          },
          "secondarySpocs": {
            "type": "array",
            "description": "List of secondary SPOCs",
            "items": {
              "$ref": "#/components/schemas/SecondarySpocDto"
            },
            "maxItems": 5
          },
          "isPinned": {
            "type": "boolean",
            "description": "Whether the drive is pinned",
            "default": false,
            "example": false
          },
          "isCompleted": {
            "type": "boolean",
            "description": "Whether the drive is completed",
            "default": false,
            "example": false
          },
          "rounds": {
            "type": "array",
            "description": "List of rounds in the drive",
            "items": {
              "$ref": "#/components/schemas/RoundDto"
            },
            "maxItems": 5
          }
        }
      },
      "UpdateDriveDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the drive",
            "example": "Updated Summer 2025 Recruitment Drive"
          },
          "collegeId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the college",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "collegeName": {
            "type": "string",
            "description": "Name of the college",
            "example": "Harvard University"
          },
          "role": {
            "type": "string",
            "enum": ["Associate Engineer", "Business Analyst"],
            "description": "Role being recruited for",
            "example": "Associate Engineer"
          },
          "practice": {
            "type": "string",
            "enum": ["Application Development", "DevOps", "PMO", "BaUX"],
            "description": "Practice area",
            "example": "Application Development"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Start date of the drive",
            "example": "2025-06-15T09:00:00Z"
          },
          "primarySpocId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the primary SPOC",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "primarySpocEmail": {
            "type": "string",
            "format": "email",
            "description": "Email of the primary SPOC",
            "example": "primary.spoc@example.com"
          },
          "primarySpocName": {
            "type": "string",
            "description": "Name of the primary SPOC",
            "example": "John Doe"
          },
          "secondarySpocs": {
            "type": "array",
            "description": "List of secondary SPOCs",
            "items": {
              "$ref": "#/components/schemas/SecondarySpocDto"
            },
            "maxItems": 5
          },
          "isPinned": {
            "type": "boolean",
            "description": "Whether the drive is pinned",
            "example": true
          },
          "isCompleted": {
            "type": "boolean",
            "description": "Whether the drive is completed",
            "example": false
          },
          "rounds": {
            "type": "array",
            "description": "List of rounds in the drive",
            "items": {
              "$ref": "#/components/schemas/RoundDto"
            },
            "maxItems": 5
          }
        }
      },
      "DriveDocument": {
        "type": "object",
        "properties": {
          "driveId": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the drive",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Name of the drive",
            "example": "Summer 2025 Recruitment Drive"
          },
          "collegeId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the college",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "collegeName": {
            "type": "string",
            "description": "Name of the college",
            "example": "Harvard University"
          },
          "role": {
            "type": "string",
            "enum": ["Associate Engineer", "Business Analyst"],
            "description": "Role being recruited for",
            "example": "Associate Engineer"
          },
          "practice": {
            "type": "string",
            "enum": ["Application Development", "DevOps", "PMO", "BaUX"],
            "description": "Practice area",
            "example": "Application Development"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Start date of the drive",
            "example": "2025-06-15T09:00:00Z"
          },
          "primarySpocId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the primary SPOC",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "primarySpocEmail": {
            "type": "string",
            "format": "email",
            "description": "Email of the primary SPOC",
            "example": "primary.spoc@example.com"
          },
          "primarySpocName": {
            "type": "string",
            "description": "Name of the primary SPOC",
            "example": "John Doe"
          },
          "secondarySpocs": {
            "type": "array",
            "description": "List of secondary SPOCs",
            "items": {
              "$ref": "#/components/schemas/SecondarySpocDto"
            }
          },
          "isPinned": {
            "type": "boolean",
            "description": "Whether the drive is pinned",
            "example": false
          },
          "isCompleted": {
            "type": "boolean",
            "description": "Whether the drive is completed",
            "example": false
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the drive is active",
            "example": true
          },
          "wecpTestIds": {
            "type": "array",
            "description": "List of WECP test IDs",
            "items": {
              "type": "string"
            }
          },
          "rounds": {
            "type": "array",
            "description": "List of rounds in the drive",
            "items": {
              "$ref": "#/components/schemas/RoundDto"
            }
          },
          "createdTimestamp": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the drive was created",
            "example": "2025-03-25T07:00:00.000Z"
          },
          "updatedTimestamp": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the drive was last updated",
            "example": "2025-03-25T07:00:00.000Z"
          }
        }
      },
      "PaginatedDriveResponse": {
        "type": "object",
        "properties": {
          "drives": {
            "type": "array",
            "description": "List of drives",
            "items": {
              "$ref": "#/components/schemas/DriveDocument"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total number of drives",
            "example": 100
          },
          "page": {
            "type": "integer",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "description": "Number of items per page",
            "example": 10
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/drives": {
      "post": {
        "tags": ["drives"],
        "summary": "Create a new drive",
        "description": "Create a new recruitment drive with the provided details",
        "operationId": "createDrive",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriveDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The drive has been successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriveDocument"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          }
        }
      },
      "get": {
        "tags": ["drives"],
        "summary": "Get a paginated list of drives",
        "description": "Get a paginated list of drives with optional filtering",
        "operationId": "getDrives",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (starts from 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of drives retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedDriveResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          }
        }
      }
    },
    "/drives/{driveId}": {
      "get": {
        "tags": ["drives"],
        "summary": "Get a drive by ID",
        "description": "Get detailed information about a specific drive by its ID",
        "operationId": "getDriveById",
        "parameters": [
          {
            "name": "driveId",
            "in": "path",
            "description": "ID of the drive to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "includeRounds",
            "in": "query",
            "description": "Whether to include rounds in the response",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Drive retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriveDocument"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          },
          "404": {
            "description": "Drive not found"
          }
        }
      },
      "put": {
        "tags": ["drives"],
        "summary": "Update a drive",
        "description": "Update an existing drive's information",
        "operationId": "updateDrive",
        "parameters": [
          {
            "name": "driveId",
            "in": "path",
            "description": "ID of the drive to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Drive updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriveDocument"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          },
          "404": {
            "description": "Drive not found"
          }
        }
      },
      "delete": {
        "tags": ["drives"],
        "summary": "Delete a drive",
        "description": "Soft delete a drive by its ID",
        "operationId": "deleteDrive",
        "parameters": [
          {
            "name": "driveId",
            "in": "path",
            "description": "ID of the drive to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Drive deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          },
          "404": {
            "description": "Drive not found"
          }
        }
      }
    },
    "/drives/{driveId}/rounds": {
      "get": {
        "tags": ["drives"],
        "summary": "Get all rounds for a drive",
        "description": "Get all rounds configured for a specific drive",
        "operationId": "getRounds",
        "parameters": [
          {
            "name": "driveId",
            "in": "path",
            "description": "ID of the drive",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "roundNumber",
            "in": "query",
            "description": "Optional round number to filter by",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rounds retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoundDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          },
          "404": {
            "description": "Drive not found"
          }
        }
      },
      "post": {
        "tags": ["drives"],
        "summary": "Create a new round",
        "description": "Create a new round for a specific drive",
        "operationId": "createRound",
        "parameters": [
          {
            "name": "driveId",
            "in": "path",
            "description": "ID of the drive",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Round created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriveDocument"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          },
          "404": {
            "description": "Drive not found"
          }
        }
      }
    },
    "/drives/{driveId}/rounds/{roundNumber}": {
      "get": {
        "tags": ["drives"],
        "summary": "Get a specific round",
        "description": "Get a specific round by number for a drive",
        "operationId": "getRoundByNumber",
        "parameters": [
          {
            "name": "driveId",
            "in": "path",
            "description": "ID of the drive",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "roundNumber",
            "in": "path",
            "description": "Round number to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Round retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoundDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          },
          "404": {
            "description": "Drive or round not found"
          }
        }
      },
      "put": {
        "tags": ["drives"],
        "summary": "Update a round",
        "description": "Update a specific round for a drive",
        "operationId": "updateRound",
        "parameters": [
          {
            "name": "driveId",
            "in": "path",
            "description": "ID of the drive",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "roundNumber",
            "in": "path",
            "description": "Round number to update",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoundDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Round updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriveDocument"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          },
          "404": {
            "description": "Drive or round not found"
          }
        }
      },
      "delete": {
        "tags": ["drives"],
        "summary": "Delete a round",
        "description": "Delete a specific round from a drive",
        "operationId": "deleteRound",
        "parameters": [
          {
            "name": "driveId",
            "in": "path",
            "description": "ID of the drive",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "roundNumber",
            "in": "path",
            "description": "Round number to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Round deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriveDocument"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          },
          "404": {
            "description": "Drive or round not found"
          }
        }
      }
    },
...
  }
}